Bugfix for DVR-MS files
Wouldn't process any DVR-MS files on windows. Bug traced back to
changing from cpp to c in 8980e807a8754b8450a4648d089a13692ff46460.

Reverting this was OK for the version of 20 May
(f7c0000add5b817e9d401ef130c28f3947a78cfe) with an own project, but not
for the current master.

Hence the change in the vcxproj, where a switch was causing this
behaviour.