Respect `-stdout` if multiple CC tracks are found
When passed the `-stdout` flag, CCExtractor should write the
subtitles to standard output, instead of an output file.

However, as noted in Issue #1453, CCExtractor doesn't
respect the `-stdout` flag when multiple CC tracks are present in
a Matroska input file (usually .mkv).

This commit ensures that output is written to standard output if `-
stdout` is present even if the input file is a Matroska container
with multiple CC tracks.

Signed-off-by: Abhishek Kumar <abhi.kr.2100@gmail.com>